+++ app/components/-private/svelte-content.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-cli-svelte/components/-private/svelte-content';