+++ tests/dummy/app/components/hello.svelte
@@ -4,8 +4,6 @@
   export let name;
 </script>
 
-<svelte:options tag="hello-ember" />
-
 <h2>Hello, {name}!</h2>
 
 {#if showSlot}