+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Default slot support (passing a block from Glimmer to Svelte)
+### Removed
+- Requirement for a wrapping DOM element
 
 ## [0.0.1] - 2022-06-26
 ### Added